Moved MessageClass enum into Message.
